requirements.txt
ussd_airflow==0.0.6
python-dotenv==0.17.0
psycopg2-binary==2.7.4